% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{assign_secret}
\alias{assign_secret}
\title{Assign a Secret Value}
\usage{
assign_secret(name, value)
}
\arguments{
\item{name}{Name of the secret to set}

\item{value}{Value to assign}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Assign a Secret Value
}
\keyword{internal}
