% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaFieldMapping.R
\name{calculate_levenshtein}
\alias{calculate_levenshtein}
\title{Calculate Levenshtein distance similarity between two strings}
\usage{
calculate_levenshtein(str1, str2)
}
\arguments{
\item{str1}{First string}

\item{str2}{Second string}
}
\value{
Numeric similarity score (0-1)
}
\description{
Alternative similarity metric using edit distance.
}
