% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{calculateResourceParams}
\alias{calculateResourceParams}
\title{Calculate optimal resource parameters}
\usage{
calculateResourceParams(total_records, mem_info, num_cores)
}
\arguments{
\item{total_records}{Total number of records to process}

\item{mem_info}{Memory information structure}

\item{num_cores}{Number of CPU cores}
}
\value{
List containing optimal chunk size and number of workers
}
\description{
Calculate optimal resource parameters
}
