% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaFieldMapping.R
\name{check_value_range_compatibility}
\alias{check_value_range_compatibility}
\title{Check if two value ranges are compatible for renaming}
\usage{
check_value_range_compatibility(
  source_range,
  target_range,
  source_field,
  target_field,
  verbose = FALSE
)
}
\arguments{
\item{source_range}{Value range of source field}

\item{target_range}{Value range of target field}

\item{source_field}{Name of source field}

\item{target_field}{Name of target field}

\item{verbose}{Whether to print verbose output}
}
\value{
List with compatibility check results
}
\description{
Check if two value ranges are compatible for renaming
}
