% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRequest.R
\name{clean}
\alias{clean}
\title{Alias for 'dataRequest'}
\usage{
clean(..., csv = FALSE, rdata = FALSE, spss = FALSE)
}
\arguments{
\item{...}{Strings, specifying the measures to process, which can be a Mongo collection, REDCap instrument, or Qualtrics survey.}

\item{csv}{Optional; Boolean, if TRUE creates a .csv extract in ./tmp.}

\item{rdata}{Optional; Boolean, if TRUE creates an .rdata extract in ./tmp.}

\item{spss}{Optional; Boolean, if TRUE creates a .sav extract in ./tmp.}
}
\value{
Prints the time taken for the data request process.
}
\description{
This is a legacy alias for the 'dataRequest' function to maintain compatibility with older code.
}
\examples{
\dontrun{
instrument_dict <- redcap_dict()
}
}
