% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{completeLoadingAnimation}
\alias{completeLoadingAnimation}
\title{Complete the loading animation}
\usage{
completeLoadingAnimation(pb)
}
\arguments{
\item{pb}{Loading animation object}
}
\description{
Complete the loading animation
}
