% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{Connect}
\alias{Connect}
\title{Setup MongoDB connection with suppressed messages}
\usage{
Connect(collection_name, db_name)
}
\arguments{
\item{collection_name}{The name of the collection you want to connect to.}

\item{db_dname}{The name of the database you cant to connect to.}
}
\value{
A mongolite::mongo object representing the connection to the MongoDB collection.
}
\description{
Setup MongoDB connection with suppressed messages
}
