% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{dataHarmonization}
\alias{dataHarmonization}
\title{Data Harmonization Function}
\usage{
dataHarmonization(df, identifier, qualtrics_alias)
}
\arguments{
\item{df}{A data frame containing the data to be harmonized.}

\item{identifier}{A string that specifies the unique identifier for the dataset;
it influences how date conversions and subsetting are handled.}

\item{collection_name}{A string representing the specific collection that needs harmonization.}
}
\value{
A data frame with the harmonized data, including standardized 'visit' column entries,
converted interview dates, and added 'measure' column based on the task.
}
\description{
This function performs data cleaning and preparation tasks, including handling missing values,
converting date formats, and adding necessary columns. It is tailored for a specific dataset
structure used in psychological or medical research.
}
\examples{
# Assuming 'df' is your dataset, 'src_subject_id' is your identifier, and 'rgpts' is your task:
harmonized_data <- dataHarmonization(df, 'src_subject_id', 'task1')

}
