% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaValidationHelpers.R
\name{detect_new_fields}
\alias{detect_new_fields}
\title{Detect fields in dataframe not in NDA structure}
\usage{
detect_new_fields(df, elements)
}
\arguments{
\item{df}{Data frame}

\item{elements}{NDA structure dataElements}
}
\value{
Character vector of new field names
}
\description{
Detect fields in dataframe not in NDA structure
}
