% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scry.R
\name{display_tree}
\alias{display_tree}
\title{Display a file tree structure similar to the Unix tree command}
\usage{
display_tree(path)
}
\arguments{
\item{path}{The path to display as a tree}
}
\value{
NULL (called for side effects)
}
\description{
Display a file tree structure similar to the Unix tree command
}
\keyword{internal}
