% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNdaDataDefinition.R
\name{exportDataDefinition}
\alias{exportDataDefinition}
\title{Export Data Definition}
\usage{
exportDataDefinition(data_definition, format = "csv")
}
\arguments{
\item{data_definition}{List containing the data definition}

\item{format}{Character string specifying export format ("json", "csv", "yaml")}
}
\description{
Exports the data definition to various formats
}
