% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaFieldMapping.R
\name{fetch_nda_structure}
\alias{fetch_nda_structure}
\title{Fetch NDA structure from API}
\usage{
fetch_nda_structure(structure_name, nda_base_url)
}
\arguments{
\item{structure_name}{Structure short name}

\item{nda_base_url}{NDA API base URL}
}
\value{
List with structure definition
}
\description{
Fetch NDA structure from API
}
