% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R, R/getRedcap.R
\name{formatDuration}
\alias{formatDuration}
\title{Format a time duration in a human-readable way}
\usage{
formatDuration(duration)

formatDuration(duration)
}
\arguments{
\item{duration}{The duration to format in seconds or minutes}
}
\value{
A formatted string representing the duration

A formatted string representing the duration
}
\description{
Format a time duration in a human-readable way

Format a time duration in a human-readable way
}
