% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaFieldMapping.R
\name{get_field_value_range}
\alias{get_field_value_range}
\title{Get field value range from NDA elements}
\usage{
get_field_value_range(field_name, elements)
}
\arguments{
\item{field_name}{Name of the field}

\item{elements}{NDA data elements}
}
\value{
Value range string or "No range specified"
}
\description{
Get field value range from NDA elements
}
