% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{getAvailableMemory}
\alias{getAvailableMemory}
\title{Cross-platform memory check function}
\usage{
getAvailableMemory()
}
\value{
List containing total and available memory in GB
}
\description{
Cross-platform memory check function
}
