% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{getMongo}
\alias{getMongo}
\title{Retrieve data from MongoDB}
\usage{
getMongo(collection_name, db_name = NULL, identifier = NULL, chunk_size = NULL)
}
\arguments{
\item{collection_name}{The name of the MongoDB collection}

\item{db_name}{The database name (optional)}

\item{identifier}{Field to use as identifier (optional)}

\item{chunk_size}{Number of records per chunk (optional)}
}
\value{
A data frame containing the MongoDB data
}
\description{
Retrieve data from MongoDB
}
