% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{getMongoData}
\alias{getMongoData}
\title{Retrieve Task Data}
\usage{
getMongoData(Mongo, identifier, batch_info)
}
\arguments{
\item{Mongo}{The MongoDB connection object.}

\item{identifier}{The document field to check for existence and non-emptiness.}

\item{batch_info}{List containing 'start' and 'size' defining the batch to fetch.}
}
\value{
A data.frame with the filtered data or NULL if no valid data is found or in case of error.
}
\description{
Retrieves data from MongoDB based on the specified batch information and query criteria.
It filters out entries where the specified identifier doesn't exist or is empty.
}
\examples{
df <- getMongoData("task_name", list(start = 0, size = 100), Mongo, "src_subject_id")
}
