% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualtrics.R
\name{getQualtrics}
\alias{getQualtrics}
\title{Retrieve Survey Data from Qualtrics}
\usage{
getQualtrics(qualtrics_alias, institution = NULL, label = FALSE)
}
\arguments{
\item{qualtrics_alias}{The alias for the Qualtrics survey to be retrieved.}

\item{institution}{Optional. The institution name (e.g., "temple" or "nu"). If NULL, all institutions will be searched.}

\item{label}{Logical indicating whether to return coded values or their associated labels (default is FALSE).}
}
\value{
A cleaned and harmonized data frame containing the survey data.
}
\description{
Retrieve Survey Data from Qualtrics
}
\examples{
\dontrun{
# Get survey by alias (will search all institutions)
survey_data <- getQualtrics("rgpts")
}
}
