% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{getRedcapForms}
\alias{getRedcapForms}
\title{Get Available REDCap Forms}
\usage{
getRedcapForms()
}
\value{
A formatted table (kable) of available REDCap instruments/forms
}
\description{
Retrieves a list of all available REDCap forms as a formatted table
}
