% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{initializeLoadingAnimation}
\alias{initializeLoadingAnimation}
\title{Initialize a clean loading animation}
\usage{
initializeLoadingAnimation(steps)
}
\arguments{
\item{steps}{Number of steps in the process}
}
\value{
Loading animation object
}
\description{
Initialize a clean loading animation
}
