% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaFileCreation.R
\name{ndaFileCreation}
\alias{ndaFileCreation}
\title{NDA File Creation Helper Functions}
\description{
Functions for creating NDA submission templates and data definition files
based on ValidationState results.
}
