% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaTransformations.R
\name{ndaTransformations}
\alias{ndaTransformations}
\title{NDA Data Transformation Functions}
\description{
Functions for transforming data to meet NDA requirements.
Includes date/age standardization, type conversions, array handling,
and missing data code translation.
}
