% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaValidator.R
\name{ndaValidator}
\alias{ndaValidator}
\title{NDA Data Structure Validator}
\usage{
ndaValidator(
  measure_name,
  api,
  limited_dataset = FALSE,
  nda_base_url = "https://nda.nih.gov/api/datadictionary/v2",
  verbose = TRUE,
  debug = FALSE,
  auto_drop_unknown = FALSE,
  interactive_mode = TRUE,
  modified_structure = NULL
)
}
\arguments{
\item{measure_name}{Name of measure/structure}

\item{api}{API type (redcap, qualtrics, mongo, csv, oracle, sql)}

\item{limited_dataset}{Logical - if TRUE, skip de-identification (date-shifting, age-capping)}

\item{nda_base_url}{NDA API base URL}

\item{verbose}{Logical - print detailed output}

\item{debug}{Logical - print debug information}

\item{auto_drop_unknown}{Logical - automatically drop unknown fields}

\item{interactive_mode}{Logical - allow user prompts}

\item{modified_structure}{Pre-enhanced NDA structure (from ndaRequest.R)}
}
\value{
ValidationState object with validation results
}
\description{
Main validation function for NDA data structure compliance.
This is the refactored version using modular components.
}
