% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOracle.R
\name{oracle.desc}
\alias{oracle.desc}
\title{Get Oracle table columns/metadata}
\usage{
oracle.desc(table_name, schema = NULL)
}
\arguments{
\item{table_name}{Name of the table to get metadata for}

\item{schema}{Optional schema name}
}
\value{
A data frame with column information
}
\description{
Get Oracle table columns/metadata
}
