% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualtrics.R
\name{qualtrics.dict}
\alias{qualtrics.dict}
\title{Fetch Qualtrics survey metadata to be stored in data frame}
\usage{
qualtrics.dict(survey_alias, exclude_embedded = TRUE)
}
\arguments{
\item{survey_alias}{Can either be an existing dataframe, variable name as string, or survey alias string}

\item{exclude_embedded}{Only select QIDs}
}
\value{
A list containing the mappings of column names to survey questions.
}
\description{
This function extracts column mappings from the metadata of a Qualtrics survey data frame.
It can accept either a data frame containing Qualtrics data, a variable name as string,
or a survey alias string.
}
