% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualtrics.R
\name{qualtrics.index}
\alias{qualtrics.index}
\title{Display table of available Qualtrics surveys}
\usage{
qualtrics.index(institution = NULL)
}
\arguments{
\item{institution}{Optional; the institution identifier to use. If NULL, uses all
institutions specified in the configuration file.}
}
\value{
A data frame containing the IDs and names of all available surveys
in the configured Qualtrics account. Can be used to identify surveys for
further data retrieval.
}
\description{
Retrieves a list of all available surveys in the configured Qualtrics account.
}
