% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{redcap.codex}
\alias{redcap.codex}
\title{Alias for 'getRedcapDictionary'}
\usage{
redcap.codex(redcap_data)
}
\arguments{
\item{redcap_data}{Can either be an instrument name to fetch new data, a data frame
returned by getRedcap(), or a variable name as string}
}
\value{
A data frame containing the data dictionary/metadata for the specified instrument
}
\description{
This is a legacy alias for the 'getRedcapDictionary' function to maintain compatibility with older code.
}
\examples{
\dontrun{
instrument_dict <- redcap.codex()
}
}
