% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{redcap.dict}
\alias{redcap.dict}
\title{Fetch REDCap data dictionary to be stored in data frame}
\usage{
redcap.dict(instrument_name)
}
\arguments{
\item{instrument_name}{Can either be an instrument name to fetch new data, a data frame
returned by redcap(), or a variable name as string}
}
\value{
A data frame containing the data dictionary/metadata for the specified instrument
}
\description{
This function extracts metadata/dictionary information from REDCap. It can accept
either an instrument name to fetch new data, an existing data frame with instrument
attributes, or a variable name as string.
}
