% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{redcap.index}
\alias{redcap.index}
\title{Display table of available REDCap instruments and their labels}
\usage{
redcap.index()
}
\value{
A formatted table (kable) of available REDCap instruments/forms
}
\description{
Retrieves a list of all available REDCap forms as a formatted table
}
