% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{show_loading_animation_with_chunks}
\alias{show_loading_animation_with_chunks}
\title{Initialize or Update Loading Animation}
\usage{
show_loading_animation_with_chunks(step = NULL, total_steps = NULL, pb = NULL)
}
\arguments{
\item{step}{Current step (chunk being processed).}

\item{total_steps}{Total number of chunks.}

\item{pb}{An existing txtProgressBar object; if NULL, a new progress bar is created.}
}
\value{
A txtProgressBar object representing the current state of the progress bar.
}
\description{
Initializes a new loading animation or updates an existing one based on the progress.
}
\examples{
pb <- show_loading_animation(step = 1, total_steps = 20)  # Start new animation
pb <- show_loading_animation(step = 2, total_steps = 20, pb = pb)  # Update animation
}
