% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSql.R
\name{sql.desc}
\alias{sql.desc}
\title{Get SQL table columns/metadata}
\usage{
sql.desc(table_name)
}
\arguments{
\item{table_name}{Name of the table to get metadata for}
}
\value{
A data frame with column information
}
\description{
Get SQL table columns/metadata
}
