% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSql.R
\name{sql.index}
\alias{sql.index}
\title{Get a list of tables from the SQL database}
\usage{
sql.index(pattern = NULL, schema = NULL)
}
\arguments{
\item{pattern}{Optional regex string; if supplied, only tables whose name
matches (case-insensitive) are shown.}

\item{schema}{Optional schema name to filter tables}
}
\value{
A data frame with table information
}
\description{
Get a list of tables from the SQL database
}
