% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSql.R
\name{sql.tables}
\alias{sql.tables}
\title{Get a list of tables from the SQL database}
\usage{
sql.tables(schema = NULL)
}
\arguments{
\item{schema}{Optional schema name to filter tables}
}
\value{
A data frame with table information
}
\description{
Get a list of tables from the SQL database
}
