% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaTransformations.R
\name{standardize_age}
\alias{standardize_age}
\title{Standardize interview_age for NDA submission}
\usage{
standardize_age(df, verbose = TRUE, limited_dataset = limited_dataset)
}
\arguments{
\item{df}{Data frame}

\item{verbose}{Logical - print detailed output}

\item{limited_dataset}{Logical - if FALSE, performs age-capping}
}
\value{
Data frame with standardized age
}
\description{
Caps age at 1068 months (89 years) for de-identification when limited_dataset=FALSE.
NDA requires age-capping for privacy protection.
}
