% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaTransformations.R
\name{standardize_column_names}
\alias{standardize_column_names}
\title{Standardize column names}
\usage{
standardize_column_names(df, structure_name, verbose = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{structure_name}{NDA structure name}

\item{verbose}{Logical - print changes}
}
\value{
Data frame with standardized column names
}
\description{
Standardizes column names by replacing hyphens with underscores and
ensuring consistent prefix_number format.
}
