% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaTransformations.R
\name{standardize_field_names}
\alias{standardize_field_names}
\title{Generic function for field standardization}
\usage{
standardize_field_names(df, measure_name, verbose = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{measure_name}{Name of measure}

\item{verbose}{Logical - print transformations}
}
\value{
Data frame with standardized field names
}
\description{
Standardizes common field names (e.g., "index" -> "trial" for behavioral tasks)
}
