% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaTransformations.R
\name{standardize_handedness}
\alias{standardize_handedness}
\title{Standardize handedness values to NDA format}
\usage{
standardize_handedness(value)
}
\arguments{
\item{value}{Character vector of handedness values}
}
\value{
Character vector with standardized values
}
\description{
Converts handedness values to standard codes: L (left), R (right), B (both/ambidextrous)
}
