% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRda.R
\name{to.rda}
\alias{to.rda}
\title{Alias for 'createRda'}
\usage{
to.rda(df, df_name = NULL)
}
\arguments{
\item{df}{DataFrame to be exported to RDS format.}

\item{df_name}{Optional; a custom file name for the saved RDS file.
If not provided, the name of the DataFrame variable is used.
The function adds the ".Rda" extension automatically.}
}
\value{
The function writes an RDS file to the specified path and prints a message
indicating the file's location. This function does not return a value.
}
\description{
This is a legacy alias for the 'createRda' function to maintain compatibility with older code.
}
\examples{
\dontrun{
to.rda(prl01)
}
}
