% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{updateLoadingAnimation}
\alias{updateLoadingAnimation}
\title{Update the loading animation}
\usage{
updateLoadingAnimation(pb, current)
}
\arguments{
\item{pb}{Loading animation object}

\item{current}{Current step}
}
\description{
Update the loading animation
}
