% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecretsEnv.R
\name{validate_secrets}
\alias{validate_secrets}
\title{Create a wrapper function to make validation easier}
\usage{
validate_secrets(api_type = NULL, secrets_file = "secrets.R")
}
\arguments{
\item{api_type}{API type to validate}

\item{secrets_file}{Path to secrets file}
}
\value{
TRUE if validation passes
}
\description{
Create a wrapper function to make validation easier
}
