% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{wizaRdry-deprecated}
\alias{wizaRdry-deprecated}
\title{Deprecated functions in wizaRdry}
\description{
Deprecated functions in wizaRdry
}
\details{
These functions are deprecated and may be removed in a future release.
Prefer the suggested replacements.
}
\section{Deprecated functions}{

\describe{
\item{\code{createCsv(...)}}{Use \code{to.csv(...)} instead.}
\item{\code{createRds(...)}}{Use \code{to.rds(...)} instead.}
\item{\code{createSpss(...)}}{Use \code{to.sav(...)} instead.}
\item{\code{dataFilter(...)}}{Use \code{sift(...)} instead.}
\item{\code{dataMerge(...)}}{Use \code{meld(...)} instead.}
\item{\code{dataRequest(...)}}{Use \code{clean(...)} instead.}
\item{\code{getRedcap(...)}}{Use \code{redcap(...)} instead.}
\item{\code{getSurvey(...)}}{Use \code{qualtrics(...)} instead.}
\item{\code{getTask(...)}}{Use \code{mongo(...)} instead.}
\item{\code{ndaRequest(...)}}{Use \code{nda(...)} instead.}
}
}

\seealso{
help("Deprecated")
}
