% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRda.R
\name{createRda}
\alias{createRda}
\title{Create RDS File from a DataFrame}
\usage{
createRda(df, df_name = NULL)
}
\arguments{
\item{df}{DataFrame to be exported to RDS format.}

\item{df_name}{Optional; a custom file name for the saved RDS file.
If not provided, the name of the DataFrame variable is used.
The function adds the ".Rda" extension automatically.}
}
\value{
The function writes an RDS file to the specified path and prints a message
indicating the file's location. This function does not return a value.
}
\description{
This function exports a given R DataFrame to an RDS file format.
The resulting file is saved in the "./tmp/" directory. If a filename
is not specified, the function uses the name of the DataFrame variable.
The ".Rda" extension is appended automatically to the filename.
}
\examples{
\dontrun{
createRds(kamin)
}
}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
