% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpss.R
\name{createSpss}
\alias{createSpss}
\title{Alias for 'to.sav'}
\usage{
createSpss(df, df_name = NULL)
}
\arguments{
\item{df}{DataFrame to be exported to SPSS format.}

\item{df_name}{Optional; custom file name for the saved SPSS file. If not provided,
the name of the DataFrame variable will be used. The ".sav" extension will
be appended automatically.}
}
\value{
Writes an SPSS file to the designated path and prints a message indicating
the file's location. This function does not return any value.
}
\description{
This is a legacy alias for the 'to.sav' function to maintain compatibility with older code.
}
\examples{
\dontrun{
createSpss(prl01)
}
}
