% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpss.R
\name{createSpss}
\alias{createSpss}
\title{Create SPSS file from a DataFrame}
\usage{
createSpss(df, df_name = NULL)
}
\arguments{
\item{df}{DataFrame to be exported to SPSS format.}

\item{df_name}{Optional; custom file name for the saved SPSS file. If not provided,
the name of the DataFrame variable will be used. The ".sav" extension will
be appended automatically.}
}
\value{
Writes an SPSS file to the designated path and prints a message indicating
the file's location. This function does not return any value.
}
\description{
This function takes a R DataFrame and writes it to an SPSS file
using the Haven package. The resulting file will be stored in the
"./tmp/" directory with a default name derived from the DataFrame variable name,
but can be customized if desired.
}
\examples{
\dontrun{
createSpss(rgpts)
}
}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
