% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMerge.R
\name{dataMerge}
\alias{dataMerge}
\title{Data Merge}
\usage{
dataMerge(
  ...,
  by = NULL,
  all = TRUE,
  no.dups = FALSE,
  csv = FALSE,
  rdata = FALSE,
  spss = FALSE
)
}
\arguments{
\item{...}{Clean data frames to be merged.}

\item{by}{A vector of strings specifying the column names to be used as merge keys. If NULL,
the function automatically determines common keys from the provided data frames.}

\item{all}{Logical; if TRUE, performs an OUTER JOIN. If FALSE, performs an INNER JOIN.}

\item{no.dups}{Logical; if TRUE, duplicates are removed post-merge.}

\item{csv}{Logical; if TRUE, the merged data frame is exported as a CSV file.}

\item{rdata}{Logical; if TRUE, the merged data frame is saved as an Rda file.}

\item{spss}{Logical; if TRUE, the merged data frame is exported as an SPSS file.}
}
\value{
A merged data frame based on the specified or common candidate keys.
}
\description{
This function simplifies the process of merging multiple cleaned data frames by automatically determining
common merge keys or utilizing user-specified keys. Supports both inner and outer join methods,
and offers options for exporting the merged data.
}
\examples{
\dontrun{
# Perform an OUTER JOIN on 'prl' and 'rgpts' using default keys:
dataMerge(prl, rgpts, all = TRUE)
# Perform an INNER JOIN using 'subjectkey' as the merge key:
dataMerge(prl, rgpts, by = "subjectkey", all = FALSE)
# Merge multiple data frames using specified keys and export to CSV:
dataMerge(prl, rgpts, lshsr, by = c("src_subject_id", "visit"), csv = TRUE)
}
}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
