% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRequest.R
\name{dataRequest}
\alias{dataRequest}
\title{Data Request}
\usage{
dataRequest(..., csv = FALSE, rdata = FALSE, spss = FALSE)
}
\arguments{
\item{...}{Strings, specifying the measures to process, which can be a Mongo collection, REDCap instrument, or Qualtrics survey.}

\item{csv}{Optional; Boolean, if TRUE creates a .csv extract in ./tmp.}

\item{rdata}{Optional; Boolean, if TRUE creates an .rdata extract in ./tmp.}

\item{spss}{Optional; Boolean, if TRUE creates a .sav extract in ./tmp.}
}
\value{
Prints the time taken for the data request process.
}
\description{
This function processes requests for clean data sequentially for specified measures.
It makes a request to the appropriate API for the named measure or measures
and runs the associated data cleaning routines. It then runs a series of
unit tests to verify that the data quality standards are met.
}
\examples{
\dontrun{
  dataRequest("prl", csv=TRUE)
  dataRequest("rgpts", "kamin", rdata=TRUE)
}

}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
