% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualtrics.R
\name{getQualtricsDictionary}
\alias{getQualtricsDictionary}
\title{Extract Column Mapping from Qualtrics Data Frame}
\usage{
getQualtricsDictionary(qualtrics_df)
}
\arguments{
\item{qualtrics_df}{Data frame obtained from Qualtrics.}
}
\value{
A list containing the mappings of column names to survey questions.
}
\description{
This function extracts column mappings from the metadata of a Qualtrics survey data frame.
}
