% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualtrics.R
\name{getSurvey}
\alias{getSurvey}
\alias{getQualtrics}
\title{Alias for 'getSurvey'}
\usage{
getSurvey(qualtrics_alias, institution = NULL, label = FALSE)
}
\description{
This is an alias for the 'getSurvey' function for ease of use.
}
\examples{
survey_data <- getQualtrics("your_survey_alias")
}
