% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{mongo.index}
\alias{mongo.index}
\title{Display table of available MongoDB collections}
\usage{
mongo.index(database = NULL)
}
\arguments{
\item{database}{Optional; the name of the database to connect to. If NULL, uses the database
specified in the configuration file.}
}
\value{
A character vector containing the names of all available collections
in the configured MongoDB database.
}
\description{
Retrieves a list of all available collections in the configured MongoDB database.
}
