% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongo.R
\name{mongo}
\alias{mongo}
\title{Retrieve data from MongoDB}
\usage{
mongo(
  collection_name,
  db_name = NULL,
  identifier = NULL,
  chunk_size = NULL,
  verbose = FALSE
)
}
\arguments{
\item{collection_name}{The name of the MongoDB collection}

\item{db_name}{The database name (optional)}

\item{identifier}{Field to use as identifier (optional)}

\item{chunk_size}{Number of records per chunk (optional)}

\item{verbose}{Logical; if TRUE, displays detailed progress messages. Default is FALSE.}
}
\value{
A data frame containing the MongoDB data with superkeys first
}
\description{
Retrieve data from MongoDB
}
\examples{
\dontrun{
# Get data from MongoDB collection
data <- mongo("collection_name")
}
}
