% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndaRequest.R
\name{ndaRequest}
\alias{ndaRequest}
\title{NDA Request}
\usage{
ndaRequest(
  ...,
  csv = FALSE,
  rdata = FALSE,
  spss = FALSE,
  limited_dataset = FALSE
)
}
\arguments{
\item{...}{Strings, specifying the measures to process, which can be a Mongo collection, REDCap instrument, or Qualtrics survey.}

\item{csv}{Optional; Boolean, if TRUE creates a .csv extract in ./tmp.}

\item{rdata}{Optional; Boolean, if TRUE creates an .rdata extract in ./tmp.}

\item{spss}{Optional; Boolean, if TRUE creates a .sav extract in ./tmp.}

\item{limited_dataset}{Optional; Boolean, if TRUE does not perform date-shifting of interview_date or age-capping of interview_age}
}
\value{
Prints the time taken for the data request process.
}
\description{
This function processes requests for clean data sequentially for specified measures.
It makes a request to the NIH NDA API for the named data structures
and runs the associated data remediation routines. It then runs a series of
unit tests to verify that the data quality standards are met.
}
\examples{
\dontrun{
  ndaRequest("prl", csv=TRUE)
  ndaRequest("rgpts", "kamin", rdata=TRUE)
}

}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
