% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOracle.R
\name{oracle.index}
\alias{oracle.index}
\title{Get a list of tables from the Oracle database}
\usage{
oracle.index(schema = NULL)
}
\arguments{
\item{schema}{Optional schema name to filter tables}
}
\value{
A data frame with table information
}
\description{
Get a list of tables from the Oracle database
}
