% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{redcap.index}
\alias{redcap.index}
\title{Alias for 'getRedcapForms'}
\usage{
redcap.index()
}
\value{
A formatted table (kable) of available REDCap instruments/forms
}
\description{
This is a legacy alias for the 'getRedcapForms' function to maintain compatibility with older code.
}
\examples{
\dontrun{
redcap.index()
}
}
