% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFilter.R
\name{sift}
\alias{sift}
\title{Data Filter}
\usage{
sift(
  df,
  rows_of_interest = NULL,
  columns_of_interest = NULL,
  visit = NULL,
  week = NULL,
  states = NULL,
  arm = NULL,
  site = NULL,
  sex = NULL,
  phenotype = NULL,
  interview_date = NULL
)
}
\arguments{
\item{df}{Dataframe to be filtered and trimmed based on the provided parameters.}

\item{rows_of_interest}{Optional; a vector of row names to be retained in the final output.
If NULL or empty, all rows in the dataframe are retained.}

\item{columns_of_interest}{Optional; a vector of column names to be retained in the final output.
If NULL or empty, all columns in the dataframe are retained.}

\item{visit}{Optional; a specific visit value to filter the dataframe by. Only used if
'visit' column exists in the dataframe.}

\item{week}{Optional; a specific week value to filter the dataframe by. Only used if
'week' column exists in the dataframe.}

\item{states}{Optional; a vector of state conditions to filter the dataframe by. Only used
if 'state' column exists in the dataframe. Can include values like 'complete',
'completed baseline', 'completed 12m', 'completed 24m', etc.}

\item{arm}{Optional; a condition specified as part of a trial e.g. drug, placebo}

\item{site}{Optional; a site like Yale, NU}

\item{sex}{Optional; a string of sex at birth 'M', 'F'}

\item{phenotype}{Optional; a string of phenotype to drill down}

\item{interview_date}{Optional; a string in MM/DD/YYYY format to lock data}
}
\value{
A filtered dataframe based on the provided 'visit', 'week', and 'state' parameters,
and containing only the columns specified in 'columns_of_interest'. If no columns
are specified, returns the entire dataframe with applied row filters.
}
\description{
Filters a dataframe based on specified 'visit', 'week', and 'state' parameters and
selects columns based on 'columns_of_interest'. If 'columns_of_interest' is not provided,
retains all columns by default. Filters by 'visit' and 'week' if these columns exist
and values for them are provided.
}
\examples{
\donttest{
# Create a sample dataframe
sample_df <- data.frame(
  src_subject_id = c("SUB001", "SUB002", "SUB003"),
  phenotype = c("A", "B", "A"),
  visit = c(1, 2, 2),
  state = c("complete", "completed baseline", "in progress"),
  stringsAsFactors = FALSE
)

# Set row names for demonstration
rownames(sample_df) <- c("foo", "bar", "baz")

# Filter the dataframe
filtered <- sift(sample_df, 
                        rows_of_interest = c("foo","bar"),
                        columns_of_interest = c("src_subject_id", "phenotype"), 
                        visit = 2, 
                        states = c("complete", "completed baseline"))
}
}
