% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSql.R
\name{sql.query}
\alias{sql.query}
\title{Perform a direct SQL query with minimal processing}
\usage{
sql.query(query, exclude_pii = FALSE)
}
\arguments{
\item{query}{The SQL query to execute}

\item{exclude_pii}{Default TRUE to remove all fields marked as identifiable}
}
\value{
A data frame with the query results
}
\description{
Perform a direct SQL query with minimal processing
}
