% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSql.R
\name{sql.query}
\alias{sql.query}
\title{Perform a direct SQL query with minimal processing}
\usage{
sql.query(query, pii = FALSE)
}
\arguments{
\item{query}{The SQL query to execute}

\item{pii}{Logical; if FALSE (default), remove fields marked as PII. TRUE keeps PII.}
}
\value{
A data frame with the query results
}
\description{
Perform a direct SQL query with minimal processing
}
