% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpss.R
\name{to.sav}
\alias{to.sav}
\title{Create SPSS file from a DataFrame}
\usage{
to.sav(df, df_name = NULL)
}
\arguments{
\item{df}{DataFrame to be exported to SPSS format.}

\item{df_name}{Optional; custom file name for the saved SPSS file. If not provided,
the name of the DataFrame variable will be used. The ".sav" extension will
be appended automatically.}
}
\value{
Writes an SPSS file to the designated path and prints a message indicating
the file's location. This function does not return any value.
}
\description{
This function takes a R DataFrame and writes it to an SPSS file
using the Haven package. The resulting file will be stored in the
"./tmp/" directory with a default name derived from the DataFrame variable name,
but can be customized if desired.
}
\examples{
\dontrun{
# Create a sample dataframe
sample_df <- data.frame(
  id = 1:3,
  score = c(85, 92, 78),
  group = c("A", "B", "A")
)
to.sav(sample_df)
}
}
\author{
Joshua Kenney \href{mailto:joshua.kenney@yale.edu}{joshua.kenney@yale.edu}
}
