% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWizardry-package.R
\docType{package}
\name{wizaRdry-package}
\alias{wizaRdry}
\alias{wizaRdry-package}
\title{wizaRdry: A Framework For Collaborative & Reproducable Data Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
