% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWizardry-package.R
\docType{package}
\name{wizaRdry-package}
\alias{wizaRdry}
\alias{wizaRdry-package}
\title{wizaRdry: A Magical Framework for Collaborative & Reproducible Data Analysis}
\description{
A comprehensive data analysis framework for NIH-funded research that streamlines workflows for both data cleaning and preparing NIH Data Archive ('NDA') submission templates. Provides unified access to multiple data sources ('REDCap', 'MongoDB', 'Qualtrics') through interfaces to their APIs, with specialized functions for data cleaning, filtering, merging, and parsing. Features automatic validation, field harmonization, and memory-aware processing to enhance reproducibility in multi-site collaborative research as described in Mittal et al. (2021) \doi{10.20900/jpbs.20210011}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/belieflab/wizaRdry}
  \item Report bugs at \url{https://github.com/belieflab/wizaRdry/issues}
}

}
\author{
\strong{Maintainer}: Joshua G. Kenney \email{joshua.kenney@yale.edu}

Authors:
\itemize{
  \item Trevor F. Williams \email{trevormsu@gmail.com}
  \item Minerva K. Pappu \email{minerva.pappu@yale.edu}
  \item Michael J. Spilka \email{michael.j.spilka@gmail.com}
}

Other contributors:
\itemize{
  \item Danielle N. Pratt \email{danielle.pratt33@gmail.com} [contributor]
  \item Victor J. Pokorny \email{victor.pokorny@northwestern.edu} [contributor]
  \item Santiago Castiello de Obeso \email{santiago.castiellodeobeso@yale.edu} [contributor]
  \item Praveen Suthaharan \email{praveen.suthaharan@yale.edu} [contributor]
  \item Christian R. Horgan \email{christian.horgan@yale.edu} [contributor]
}

}
\keyword{internal}
